+++ src/Caption.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
-import { prepareCaption } from './helpers';
-import { CaptionProps } from '../types';
+import { prepareCaption } from './helpers/prepareCaption';
+import { CaptionProps } from './types';
 export const Caption: React.FC<CaptionProps> = ({ month, dayPickerProps }) => {
   const { containerProps } = prepareCaption(dayPickerProps);
   const { locale } = dayPickerProps;
