--- src/classes/index.ts
@@ -1,1 +1,0 @@
-export { DateWithModifiers } from './DateWithModifiers';
