--- src/classes/utils/index.ts
@@ -1,3 +1,0 @@
-export { getModifiersFromProps } from './getModifiersFromProps';
-export { listModifiers } from './listModifiers';
-export { matchModifier } from './matchModifier';
