--- src/components/helpers/index.ts
@@ -1,5 +1,0 @@
-export { prepareCaption } from './prepareCaption';
-export { prepareDay } from './prepareDay';
-export { prepareDayPicker } from './prepareDayPicker';
-export { prepareMonth } from './prepareMonth';
-export { prepareNavigation } from './prepareNavigation';
