--- src/components/utils/index.ts
@@ -1,4 +1,0 @@
-export { filterUndefinedProps } from './filterUndefinedProps';
-export { getOutsideEndDays } from './getOutsideEndDays';
-export { getOutsideStartDays } from './getOutsideStartDays';
-export { getWeekdaysNames } from './getWeekdaysNames';
