+++ src/index.ts
@@ -1,2 +1,15 @@
-export * from './components';
-export * from './components/helpers';
+export { Caption } from './Caption';
+export { Day } from './Day';
+export { DayPicker } from './DayPicker';
+export { Head } from './Head';
+export { Month } from './Month';
+export { Navigation } from './Navigation';
+export { Week } from './Week';
+export { DateWithModifiers } from './DateWithModifiers';
+export { defaultClassNames } from './defaultClassNames';
+
+export { prepareDay } from './helpers/prepareDay';
+export { prepareCaption } from './helpers/prepareCaption';
+export { prepareDayPicker } from './helpers/prepareDayPicker';
+export { prepareMonth } from './helpers/prepareMonth';
+export { prepareNavigation } from './helpers/prepareNavigation';
